[DF-317] Fixed warnings for Win32
Former-commit-id: 5b83d2a8aa58460cc6d2c8ba452cf1288853aca1

+++ Tools/ResourceEditor/Classes/LandscapeEditor/NotPassableTerrain.cpp
@@ -38,7 +38,7 @@ NotPassableTerrain::NotPassableTerrain()
     SetName(String("Landscape_NotPassable"));
 
     notPassableAngleTan = (float32)tan(DegToRad((float32)NotPassableTerrain::NOT_PASSABLE_ANGLE));
-    notPassableMapSprite = Sprite::CreateAsRenderTarget(TEXTURE_TILE_FULL_SIZE, TEXTURE_TILE_FULL_SIZE, DAVA::FORMAT_RGBA8888);
+    notPassableMapSprite = Sprite::CreateAsRenderTarget((float32)TEXTURE_TILE_FULL_SIZE, (float32)TEXTURE_TILE_FULL_SIZE, DAVA::FORMAT_RGBA8888);
 }
 
 NotPassableTerrain::~NotPassableTerrain()
